Create AboutME.md
I'd not have filled anything here, but I'm learning about documentation. I also worked on with this generative code for blender from Blend My NFTs, and the dev did this extensive documentation that made me understand it so well. So from here on out, I want to commit to writing descriptions and documentations for everything I do.

How does this concern you? It probably doesn't but nonetheless. yeah!

This is my first repo... About me. I'll update it as I ascend from noob to god-level.